feat: move abc and log to core, start CLI
