+++ yafti/core/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2023 Marco Ceppi
+# SPDX-License-Identifier: Apache-2.0
