+++ yafti/screen/package/screen/package.py
@@ -4,7 +4,7 @@ from gi.repository import Adw, Gtk
 
 import yafti.share
 from yafti import events
-from yafti.abc import YaftiScreen, YaftiScreenConfig
+from yafti.core.abc import YaftiScreen, YaftiScreenConfig
 from yafti.screen.package.models import PackageConfig, PackageGroupConfig
 from yafti.screen.package.screen import PackageInstallScreen, PackagePickerScreen
 from yafti.screen.package.state import PackageScreenState
