+++ yafti/screen/title.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 from gi.repository import Adw, Gtk
 
 from yafti import events
-from yafti.abc import YaftiScreen, YaftiScreenConfig
+from yafti.core.abc import YaftiScreen, YaftiScreenConfig
 from yafti.core.registry import PLUGINS
 
 _xml = """\
